nfsdcld: don't send null client ids to the kernel
It's apparently possible for the sqlite database to get corrupted and
cause one or more rows to have null in the id column.

The knfsd fix was posted here:
https://lore.kernel.org/linux-nfs/20240903111446.659884-1-lilingfeng3@huawei.com/

nfsdcld should have a similar fix.  If we encounter a client record with
a null id, just skip it instead of sending it to the kernel.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>