nfsdctl: add support for new lockd configuration interface
This patch series adds support for the new lockd configuration interface
that should fix this RH bug:

    https://issues.redhat.com/browse/RHEL-71698

There are some other improvements here too, notably a switch to xlog.
Only lightly tested, but seems to do the right thing.

Port handling with lockd still needs more work. Currently that is
usually configured by rpc.statd. I think we need to convert it to
use netlink to configure the ports as well, when it's able.

To: Steve Dickson <steved@redhat.com>
Cc: Scott Mayhew <smayhew@redhat.com>
Cc: Yongcheng Yang <yoyang@redhat.com>
Cc: linux-nfs@vger.kernel.org
Signed-off-by: Jeff Layton <jlayton@kernel.org>

---
Changes in v2:
- properly regenerate manpages
- fix up bogus merge conflict
- add D_GENERAL xlog messages when nfsdctl starts and exits
- Link to v1: https://lore.kernel.org/r/20250109-lockd-nl-v1-0-108548ab0b6b@kernel.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20250109-lockd-nl-6272fa9e8a5d",
    "prefixes": [],
    "history": {
      "v1": [
        "20250109-lockd-nl-v1-0-108548ab0b6b@kernel.org"
      ]
    }
  }
}