nfsdctl: fix autostart
During nfsdctl autostart the nfsdctl reads the nfs.conf file and
tries to setup the listeners and start knfsd threads. However,
if we failed to start a listener, it currently ignores the error
and starts the threads anyway.

Suggesting that if we fail to start a UDP/TCP listener then do not
start threads. At the moment ignoring the failure of adding an
RDMA listener (because default config might have rdma=y set but
not RDMA-enabled interface is available and previously we did not
fail the start of knfsd in that case).

Signed-off-by: Olga Kornievskaia <okorniev@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>