nfsdctl: convert to xlog()
Convert all of the fprintf(stderr, ...) calls to xlog(L_ERROR, ...)
calls.  Change the -d option to not take an argument, and add a -s
option that will make nfsdctl log to syslog instead of stderr.

Also remove the extraneous error message in run_commandline, and add
a couple of trivial debug messages when the program starts and ends

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Steve Dickson <steved@redhat.com>