sm-notify: Do not drop privileges if running as non-root user
If sm-notify is not running root and is running
as the owner of the state directory, it is not
necessary to drop privileges.

The patch checks if the current user is the owner of the state
directory.

In the case of a ha-callout, the ha-callout will, more than likely, be running
as rpcuser - or similar.  That means the sm-notify program can not be called
from the ha-callout.

nsm_drop_privileges() is also called by statd.  However, that will fail for
a number of reasons such as not being able to create the PID file etc.  So,
this patch should not cause any issues for statd.  statd will still need to be
started as root.

Without this patch capabalities such as cap_setpcap, cap_setgid,
cap_net_bind_service=ep need to be set.  Clearly, that is an ugly hack and
best avoided.

Signed-off-by: Peter Schwenke <pschwenke@ddn.com>
Signed-off-by: Peter Schwenke <peter@bluetoad.com.au>
Signed-off-by: Steve Dickson <steved@redhat.com>